Export things to create custom ConfigStore
Currently, implementing a custom `ConfigStore` outside of `presage` is
not possible, mainly as `ContactsStore` and `StateStore` are not
publicly accessible. Futhermore, some data in `Registered` needs to be
public to be stored in the `StateStore`. This PR basically just makes
all of these needed things public.

I also thought about maybe hiding the exports behind a feature, as most
people will probably not need to create their own `ConfigStore`, but
having the option is still needed as sometimes the built-in stores may
not suffice.