Message storage
This is a initial draft for message storage. The trait behind message
storage is `MessageStore`, which is currently only has very basic
functionality, e.g. storing and retrieving messages by sender and
timestamp, listing all messages. This should definitly be expanded in
the future with e.g. associating messages with channels/groups, listing
messages from channels/groups. I don't think searching would be easily
possible as sled (the main current implementation) is basically just a
key-value store and not intended to be searchable (as far as I see).

The message serialization/deserialization is done using protobuf. I
think the additional work done here (mainly the
`InternalSerialization.proto` with function implementations) should be
upstreamed to `libsignal-service-rs`

For receiving messages, there is a new function in the `Manager`, if the
config store is also a messages store. This is only a thin wrapper
around the normal receive function.

The example cli has also been updated to include use the new function for
receiving and a argument for listing the current messages.

As already mentioned, this is still a early draft and only represents a
idea on how this could be implemented. This idea should be discussed if
this even is the right direction for this project.