Fix storing messages
Currently, message timestamps are encrypted when storing. This does not
work when retrieving messages before a timestamp (and I think currently
does not work when getting/deleting messages in general). This PR fixes
this by just removing the encryption.

Before this PR should actually be merged, it should maybe be discussed
if removing the encryption is actually a good thing to do, but as far as
I see it is the only thing possible except for every time we look for
messages of a thread, decrypt all message keys and manually sort and
iterate them.