Encrypt registration
Previously, the registration was not encrypted. As this contains (as far
as I have seen) keys of the device, it could be used to link an
unauthorized device with the stored data.

This commit encrypts the data. This will break previously used instances
of the store and may therefore require a migration, which I have not yet
investigated.