Fix storing outgoing messages (#144)
Commit c5f23d4 broke storing outgoing messages by not setting the
message thread and instead calculating it from the message, which can
not differentiate between incoming and outgoing messages. This sets the
thread for outgoing messages like done before.

This commit also stores call messages as I think these should also be
shown in the chat-history and should therefore be stored. This is
debatable though as I have no idea how Signal is handling that.