Fix duplicate message to group (#142)
Currently, when sendinging a message to a group, this message will be
sent both as a data message to linked devices and as a sync message. As
discussed in the whisperfish-chat, the recipients-list should not
contain the own UUID. As also discussed, this breaks sending to groups
with only the sender and no other persons, this has to be fixed
upstream.