Fix outgoing messages stored in wrong thread (#216)
Outgoing messages in 1-to-1 chats were stored in the note-to-self chat.
This was broken in 6125534. This will be fixed by this MR.

This also fixes the timestamp of identity verification messages, which
are inserted into the store by presage. By convention, all timestamps
used are in milliseconds, this one was in seconds. This was wrong since #206.