fix: cache sender certificate and make manager state immutable (#316)
The sender certificate field in `Registered` was not updated in all
instances of the manager after the certificate was fetched. Now the field
is cached behind a reference-counted mutex. Additionally, we store the
certificate in the store, and it is loaded on startup.

Also, make the state field in the manager immutable. This makes it
impossible to modify the state by value.