Fix listing groups in the sqlite store (#336)
The problem is the following commit: https://github.com/whisperfish/libsignal-service-rs/commit/939b8e69df7107d843b3a9a10f883a3c8651e548.
This renamed the "uuid" field of "Member" to "aci". As we store members
as JSON in the DB, this breaks listing groups if there is at least one
group which did not yet have a message since that change was introduced.
To fix this issue, I have mirrored the Member struct into presage and
added an alias for the "aci" field.