Fix sqlite profile key store querying (#341)
Previously, any calls to `profile_key` on the sqlite store returned
`None`, which lead to all messages being sent identified. This causes
issues with rate limiting. The underlying issue was that we query the
profile key from the UUID given as string, while we insert the profile
key from the UUID given as raw bytes (I think at least that is the issue,
I'm not sure how sqlx internally handles it). The fix itself is pretty
simple: Convert the `ServiceId` to a `Uuid` and query on that, instead
of converting it to a String.

Ideally (at least I think), we should change the interface of the store
to take `Uuid` (or even an `Aci` itself) instead of a `ServiceId`, as
we only seem to query profile keys based on `Aci`.