Migrate doc tests to use sqlite store (#344)
They previously used the sled store, which has been deprecated for a
while now. This commit pretty much just replaces the store in the doc
tests with the sqlite store. While implementing it, I also noticed that
running the doc tests in the `presage` crate instead of in the workspace
also leads to errors as tokio a feature (`rt-multi-thread`) is used that
is only enabled in another crate (`presage-cli`). I therefore also added
tokio separately as a dev dependency with that feature enabled. Doing
that, I also noticed we are using two separate tokio versions (at least
in the `Cargo.toml` files, they are unified by Rust later), I therefore
replaced the old version with the new one as well.