Actually encrypt the sqlite store
We previously did not encrypt the sqlite store as we did not have
libsqlite3 with bundled sqlite-cipher. #335 tried to change that, but
that was not backwards compatible, and had to be reverted in #337. This
commit now enables the encryption as already done in #335, but also
provides a migration strategy. Sadly, migration is not as easy as just
defining an sqlx migration, and instead requires creating a new encrypted
database, moving everything from the unencrypted to the encrypted store,
and finally replacing the unencrypted store with the encrypted one. The
details are non-trivial, and layed out as comments in the actual code.