Fix avatar caching (#348)
The previous methods for caching avatars broke in the switch to sqlite.
The reason for that was that for inserting an avatar into the cache
"UPDATE ... SET ... WHERE ..." was used, which does not insert a row if
there is no entry already with the corresponding key in the "WHERE",
therefore those tables always were empty. The fix simply replaces this
pattern for "INSERT OR REPLACE INTO ... VALUES ...". This is fixed both
for profile and group avatars. I don't think this pattern is used
anywhere else as far as I can see.